Change DropShadowPanel from being a user control to a templated control
